+++ mod.ts
@@ -8,3 +8,7 @@ export { default as InputCheckCombo } from './src/molecules/InputCheckCombo.tsx'
 export { default as InputTextareaCombo } from './src/molecules/InputTextareaCombo.tsx';
 export { default as InputSelectCombo } from './src/molecules/InputSelectCombo.tsx';
 export { default as Markdown } from './src/molecules/Markdown.tsx';
+
+export { default as InterObs } from './src/organisms/InterObs.tsx';
+export { default as Keyboard } from './src/organisms/Keyboard.tsx';
+// export { default as Revealer } from './src/organisms/Revealer.tsx';
